Reorganise the code
