+++ src/Connector/ItemSet.php
@@ -1,8 +1,6 @@
 <?php
 
-declare(strict_types=1);
-
-namespace Sylake\AkeneoProducerBundle\Listener;
+namespace Sylake\AkeneoProducerBundle\Connector;
 
 final class ItemSet implements ItemSetInterface
 {
